+++ lib/igt_kms.h
@@ -218,7 +218,7 @@ uint32_t kmstest_find_crtc_for_connector(int fd, drmModeRes *res,
 					 uint32_t crtc_blacklist_idx_mask);
 
 uint32_t kmstest_dumb_create(int fd, int width, int height, int bpp,
-			     unsigned *stride, unsigned *size);
+			     unsigned *stride, uint64_t *size);
 
 void *kmstest_dumb_map_buffer(int fd, uint32_t handle, uint64_t size,
 			      unsigned prot);
